requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.32
sqlalchemy-libsql==0.1.0
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==1.0.0
